+++ Vienna/FolderHeaderBar.m
@@ -46,11 +46,14 @@
 	// Note that [currentFolder messageCount] returns -1 if it hasn't cached
 	// the messages for the folder yet. This is OK.
 	if (folderCount <= 0)
+#warning 64BIT: Check formatting arguments
 		countString = [NSString stringWithFormat:NSLocalizedString(@"No messages", nil)];
 	else
 	{
+#warning 64BIT: Check formatting arguments
 		countString = [NSString stringWithFormat:NSLocalizedString(@"Infobar Format", nil), folderCount];
 		if (currentFolder && [currentFolder unreadCount] && !IsSearchFolder(currentFolder))
+#warning 64BIT: Check formatting arguments
 			countString = [countString stringByAppendingFormat:NSLocalizedString(@" (%d unread)", nil), [currentFolder unreadCount]];
 	}
 	[unreadCount setStringValue:countString];
@@ -58,7 +61,7 @@
 
 /* setFolderCount
  */
--(void)setFolderCount:(int)newCount
+-(void)setFolderCount:(NSInteger)newCount
 {
 	folderCount = newCount;
 }
@@ -66,7 +69,7 @@
 /* setCurrentFolder
  * Updates the current folder.
  */
--(void)setCurrentFolder:(Database *)db folderId:(int)folderId
+-(void)setCurrentFolder:(Database *)db folderId:(NSInteger)folderId
 {
 	Folder * folder = [db folderFromID:folderId];
 	NSMutableString * fullTitleString = [NSMutableString stringWithString:[db folderPathName:folderId]];
